moved base to backend-common
