+++ vantage6-algorithm-store/vantage6/algorithm/store/__init__.py
@@ -45,7 +45,7 @@ from vantage6.algorithm.store.globals import API_PATH
 from vantage6.algorithm.store.globals import RESOURCES, SERVER_MODULE_NAME
 
 # TODO the following are simply copies of the same files in the server - refactor
-from vantage6.common.base import Base, DatabaseSessionManager, Database
+from vantage6.backend.common.base import Base, DatabaseSessionManager, Database
 from vantage6.algorithm.store import db
 
 # TODO move server imports to common / refactor
