+++ vantage6-algorithm-store/vantage6/algorithm/store/model/__init__.py
@@ -1,5 +1,5 @@
 # pylint: disable=unused-import
-from vantage6.common.base import Base
+from vantage6.backend.common.base import Base
 from vantage6.algorithm.store.model.algorithm import Algorithm
 from vantage6.algorithm.store.model.argument import Argument
 from vantage6.algorithm.store.model.database import Database
