+++ vantage6-algorithm-store/vantage6/algorithm/store/model/database.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 from sqlalchemy import Column, String, ForeignKey, Integer
 from sqlalchemy.orm import relationship
 
-from vantage6.common.base import Base
+from vantage6.backend.common.base import Base
 
 
 class Database(Base):
