+++ vantage6-algorithm-store/vantage6/algorithm/store/model/function.py
@@ -1,7 +1,7 @@
 from sqlalchemy import Column, String, ForeignKey, Integer
 from sqlalchemy.orm import relationship
 
-from vantage6.common.base import Base
+from vantage6.backend.common.base import Base
 
 
 class Function(Base):
