+++ vantage6-server/tests_server/test_models.py
@@ -6,7 +6,7 @@ import datetime
 from sqlalchemy.exc import IntegrityError
 
 from vantage6.server.controller.fixture import load
-from vantage6.common.base import Database, DatabaseSessionManager
+from vantage6.backend.common.base import Database, DatabaseSessionManager
 from vantage6.server.globals import PACKAGE_FOLDER, APPNAME
 
 from vantage6.backend.common import session
