+++ vantage6-server/vantage6/server/__init__.py
@@ -51,7 +51,7 @@ from vantage6.common.globals import PING_INTERVAL_SECONDS
 from vantage6.backend.common.globals import HOST_URI_ENV
 from vantage6.server import db
 from vantage6.cli.context.server import ServerContext
-from vantage6.common.base import DatabaseSessionManager, Database
+from vantage6.backend.common.base import DatabaseSessionManager, Database
 from vantage6.server.resource.common.output_schema import HATEOASModelSchema
 from vantage6.server.permission import RuleNeed, PermissionManager
 from vantage6.server.globals import (
