+++ vantage6-server/vantage6/server/model/authenticatable.py
@@ -2,7 +2,7 @@ import bcrypt
 
 from sqlalchemy import Column, String, DateTime
 
-from vantage6.common.base import Base
+from vantage6.backend.common.base import Base
 
 
 class Authenticatable(Base):
