+++ vantage6-server/vantage6/server/model/role_rule_association.py
@@ -6,7 +6,7 @@ role will usually have multiple rules assigned to it.
 
 from sqlalchemy import Column, Integer, ForeignKey, Table
 
-from vantage6.common.base import Base
+from vantage6.backend.common.base import Base
 
 
 role_rule_association = Table(
