+++ vantage6-server/vantage6/server/model/rule.py
@@ -4,7 +4,7 @@ from enum import Enum as Enumerate
 from sqlalchemy import Column, Text, String, Enum
 from sqlalchemy.orm import relationship
 from sqlalchemy.orm.exc import NoResultFound
-from vantage6.common.base import Base, DatabaseSessionManager
+from vantage6.backend.common.base import Base, DatabaseSessionManager
 
 
 class Operation(str, Enumerate):
